M38 t3: AT_FDCWD is honoured in the 32-bit form real guests pass
`translate_fds`'s sentinel check tested `(v as i64) < 0`, but libc passes `int
dirfd` in w0, so AT_FDCWD (-2) arrives in x0 as 0xffff_fffe — a positive i64 —
not the sign-extended 64-bit form the check assumed. Every real guest's
AT_FDCWD was looked up as a descriptor and got EBADF (M33 Ruling 10); a bare
FdTable unit test stayed green throughout because it only ever exercised the
sign-extended form. Fix: test the low 32 bits, `(v as i32) < 0`.

The fdxlat.rs test file has its OWN copy of translate_fds's walk (a bare-table
mirror, since booting a real Box_ per test would collide with HVF's one-VM-
per-process rule under --test-threads=1); its doc comment claims to perform
"the same walk", so the identical fix was needed there too, or the rewritten
unit test would still fail the 0xfffffffe case after the real fix landed. A
new fixture (`atfdcwd_dyn`: `fstatat(AT_FDCWD, ".")`) and e2e gate assert on
the trace directly: the guest passed the 32-bit form AND the call succeeded.

Hand check: `record-dyn /bin/ls` now issues fstatat64(0xfffffffe, ...) without
EBADF (confirmed via RETRACE_TRACE=1) — the "ls: .: Bad file descriptor"
symptom this fixes is gone. It still cannot list a directory: it hits an
unrelated, pre-existing wall (getattrlistbulk, syscall 461, no arg_kinds row)
downstream, out of this task's scope.

Gate: fdxlat 8/8, atfdcwd_e2e 1/1, plus fdtable_e2e/jq_file_e2e/sysbin_e2e
green as neighbours; retrace-arch 45/45, retrace-guest 16/16, retrace-box
263/263; clippy clean.

+++ crates/retrace-box/src/lib.rs
@@ -3212,8 +3212,11 @@ impl Box_ {
         // is exactly as loud on an unenumerated number.
         for i in retrace_arch::shape_of(num, &*args).fd_operands() {
             let v = args[i];
-            // AT_FDCWD (-2) and friends are sentinels, not descriptors.
-            if (v as i64) < 0 { continue; }
+            // AT_FDCWD (-2) and friends are sentinels, not descriptors. `int fd` arrives in w0,
+            // so the sentinel is 0xffff_fffe in x0, not the sign-extended form — the low 32 bits
+            // are what carry the sign (M33 Ruling 10, fixed M38). A real descriptor never has
+            // bit 31 set.
+            if (v as i32) < 0 { continue; }
             match self.fds.host(v) {
                 Some(h) => args[i] = h as u64,
                 // Console fds 0/1/2 DO have a host mapping (identity, `FdTable::new`); `None`
